Move RootHandlers base class to FWCore/AbstractServices
