+++ IOPool/Common/BuildFile.xml
@@ -1,3 +1,4 @@
+<use name="FWCore/AbstractServices"/>
 <use name="FWCore/ServiceRegistry"/>
 <use name="FWCore/Utilities"/>
 <export>
